add and display pet api and table completed
