+++ controllers/petController.js
@@ -1,9 +1,8 @@
-const { addPetRequest } = require('../models/petModel');
+const { insertPet, getAllPets } = require('../models/petModel');
 
-const addNewPet = async (req, res) => {
-    const { pet_name, pet_description, pet_species, pet_breed, pet_age, pet_personality, pet_special_requirement, pet_image } = req.body;
-
-    const petDetails = {
+// Function to add new pet
+async function addNewPet(req, res) {
+    const {
         pet_name,
         pet_description,
         pet_species,
@@ -12,16 +11,35 @@ const addNewPet = async (req, res) => {
         pet_personality,
         pet_special_requirement,
         pet_image
-    };
+    } = req.body;
 
     try {
-        await addPetRequest(petDetails);
+        await insertPet({
+            pet_name,
+            pet_description,
+            pet_species,
+            pet_breed,
+            pet_age,
+            pet_personality,
+            pet_special_requirement,
+            pet_image
+        });
         res.redirect('/shelter');
-    } catch (error) {
-        console.error('Error adding pet:', error);
-        res.status(500).json({ error: 'Internal server error' });
+    } catch (err) {
+        console.error(err);
+        res.status(500).send('Internal Server Error');
     }
-};
+}
 
+// Function to get all pets
+async function getPetItems(req, res) {
+    try {
+        const data = await getAllPets();
+        res.json(data);
+    } catch (err) {
+        console.error(err);
+        res.status(500).send('Internal Server Error');
+    }
+}
 
-module.exports = { addNewPet };
+module.exports = { addNewPet, getPetItems };
