+++ models/petModel.js
@@ -1,35 +1,18 @@
 const { client } = require('../dbConnection.js');
+
 const petCollection = client.db("TestDB").collection("pets");
 
-const addPetRequest = async (petDetails) => {
-    try {
-        await client.connect();
-        let newPet = {
-            ...petDetails,
-            creation_date: new Date(),
-        };
+// Function to insert a new pet
+async function insertPet(petData) {
+    return await petCollection.insertOne(petData);
+}
 
-        let result = await petCollection.insertOne(newPet);
-        return result;
-    } catch (error) {
-        throw error;
-    }
-};
-
-
-// Function to get all pet and sending them into api/pet routes
-async function getPetItems(req, res) {
-    try {
-        await client.connect();
-        let data = await petCollection.find({}).toArray();
-        res.json(data);
-    } catch (err) {
-        console.error(err);
-        res.status(500).send('Internal Server Error');
-    }
+// Function to get all pets
+async function getAllPets() {
+    return await petCollection.find({}).toArray();
 }
 
 module.exports = {
-    addPetRequest,
-    getPetItems
+    insertPet,
+    getAllPets
 };
