+++ views/shelter.html
@@ -53,7 +53,8 @@
               <tr>
                 <th scope="col">#</th>
                 <th scope="col">Pet Name</th>
-                <th scope="col">Pet Description</th>
+                <th scope="col">Pet Age</th>
+                <th scope="col">Pet Species</th>
                 <th scope="col">Action</th>
               </tr>
             </thead>
@@ -177,6 +178,7 @@
   <!-- Main Script -->
   <script src="/js/indexScript.js"></script>
   <script src="/js/logoutScript.js"></script>
+  <script src="/js/shelterScript.js"></script>
 </body>
 
 </html>
